+++ RELEASE-NOTES.md
@@ -16,6 +16,8 @@
 
 * Update Solidity compiler version to 0.4.24.
 
+* Remove use of `bignumber.js` and replace with `web3.toBigNumber()` (Thanks to [barakman](https://github.com/barakman)).
+
 ## Version 1.0.4 - 2018-06-12 ##
 
 * Document owners_ address being strictly increasing, by [ripper234](https://github.com/ripper234)